requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
mysqlclient
